historynode/src/test/java/io/coti/historynode/services/AddressServiceTest.java
package io.coti.historynode.services;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.coti.basenode.data.AddressData;
import io.coti.basenode.data.TransactionData;
import io.coti.basenode.http.BaseResponse;
import io.coti.basenode.http.interfaces.IResponse;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.junit4.SpringRunner;

import java.io.IOException;

import static io.coti.basenode.http.BaseNodeHttpStringConstants.STATUS_SUCCESS;
import static org.junit.Assert.*;
import static testUtils.TestUtils.createRandomTransaction;
import static testUtils.TestUtils.generateRandomHash;
@TestPropertySource(locations = "classpath:test.properties")
@RunWith(SpringJUnit4ClassRunner.class)
@SpringBootTest
public class AddressServiceTest {

    @Autowired
    private AddressService addressService;

    @Test
    public void addressTest() throws IOException {
        AddressData addressData = new AddressData(generateRandomHash());
        ObjectMapper mapper = new ObjectMapper();
        String transactionAsJson = mapper.writeValueAsString(new AddressData(generateRandomHash())) ;
        addressService.insertAddressJson(addressData.getHash(), transactionAsJson);
        IResponse response =  addressService.getAddressByHash(addressData.getHash()).getBody();
        Assert.assertTrue(((BaseResponse)(response)).getStatus().equals(STATUS_SUCCESS));
    }
}